+++ src/threads/destroy_mutexes.c
@@ -6,17 +6,18 @@
 /*   By: rnijhuis <rnijhuis@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/01/27 13:05:37 by rnijhuis      #+#    #+#                 */
-/*   Updated: 2022/05/17 22:04:54 by rubennijhui   ########   odam.nl         */
+/*   Updated: 2022/07/08 12:41:31 by rnijhuis      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
+
 #include <stdint.h>
 
 int32_t	destroy_mutexes(t_philosopher *philos, pthread_mutex_t stop_sim_lock, \
 	pthread_mutex_t print_lock)
 {
-	uint32_t	i;
+	size_t		i;
 	uint32_t	amount_philos;
 
 	i = 0;
